+++ lib/qc/modules/device_details/widgets/device_details_widget.dart
@@ -1,10 +1,14 @@
+import 'dart:async';
+
 import 'package:core/core.dart';
 import 'package:core_widgets/core_widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_trc/qc/modules/device_details/resources/device_detail_service.dart';
 import 'package:flutter_trc/src/common/utils/csh_ml_scanner_util.dart';
+import 'package:flutter_trc/src/utils/dotted_divider_line.dart';
 
 import '../l10n.dart';
+import '../resources/device_detail_response.dart';
 
 class DeviceDetailsWidget extends StatefulWidget {
   final String deviceBarcode;
@@ -57,21 +61,41 @@ class _DeviceDetailsWidgetState extends State<DeviceDetailsWidget> {
 
         if (snapshot.hasData) {
           var data = snapshot.data;
+          List<Widget> list = _getWidgetList(data!, theme, l10n);
           return SingleChildScrollView(
             padding: const EdgeInsets.all(Dimens.space_16),
             child: Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
               children: [
-                if (!Validator.isNullOrEmpty(data?.barcode)) _row(l10n.deviceBarcode, data!.barcode!, theme),
-                if (!Validator.isNullOrEmpty(data?.modelName)) _row(l10n.model, data!.modelName!, theme),
-                if (!Validator.isNullOrEmpty(data?.imei)) _row(l10n.imei, data!.imei!, theme),
-                if (!Validator.isNullOrEmpty(data?.serialNo)) _row(l10n.serialNo, data!.serialNo!, theme),
-                if (!Validator.isNullOrEmpty(data?.location)) _row(l10n.storageLocation, data!.location!, theme),
-                if (!Validator.isNullOrEmpty(data?.status)) _row(l10n.currentStatus, data!.status!, theme),
-                if (!Validator.isNullOrEmpty(data?.repairStatus)) _row(l10n.repairStatus, data!.repairStatus!, theme),
-                if (!Validator.isListNullOrEmpty(data?.channelList))
-                  _row(l10n.channelName, data!.channelList!.join(" | "), theme),
-                if (data?.stockAge != null) _row(l10n.stockAge, data!.stockAge.toString(), theme),
-                if (!Validator.isNullOrEmpty(data?.lotName)) _row(l10n.lotName, data!.lotName.toString(), theme),
+                CshCard(
+                  padding: EdgeInsets.all(Dimens.space_16),
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Row(
+                        children: [
+                          Expanded(child: Text(data.modelName ?? "", style: theme.primaryTextTheme.headlineMedium)),
+                          _column(l10n.barcode, data.barcode ?? "", theme),
+                        ],
+                      ),
+                      SizedBox(height: Dimens.space_8),
+                      DottedLineDivider(dashWidth: Dimens.space_4),
+                      SizedBox(height: Dimens.space_8),
+                      ...list,
+                      FutureBuilder(
+                        future: _getStockMovement(_deviceBarcode),
+                        builder: (context, snapshot) {
+                          if (snapshot.connectionState == ConnectionState.done) {
+                            if (snapshot.hasData) {
+                              return Stepper(steps: snapshot.data!);
+                            }
+                          }
+                          return SizedBox.shrink();
+                        },
+                      ),
+                    ],
+                  ),
+                ),
                 const SizedBox(height: Dimens.space_16),
                 CshMediumButton(
                     text: l10n.scanOtherDevice,
@@ -87,6 +111,64 @@ class _DeviceDetailsWidgetState extends State<DeviceDetailsWidget> {
     );
   }
 
+  List<Widget> _getWidgetList(DeviceDetailResponse detail, ThemeData theme, L10n l10n) {
+    List<Widget> widgetList = [];
+    if (!Validator.isNullOrEmpty(detail.imei)) {
+      widgetList.add(_column(l10n.imei1, detail.imei!, theme));
+    }
+    if (!Validator.isNullOrEmpty(detail.imei2)) {
+      widgetList.add(_column(l10n.imei2, detail.imei2!, theme));
+    }
+    if (!Validator.isNullOrEmpty(detail.serialNo)) {
+      widgetList.add(_column(l10n.serialNo, detail.serialNo!, theme));
+    }
+    if (!Validator.isNullOrEmpty(detail.location)) {
+      widgetList.add(_column(l10n.location, detail.location!, theme));
+    }
+    if (!Validator.isNullOrEmpty(detail.status)) {
+      widgetList.add(_column(l10n.currentStatus, detail.status!, theme));
+    }
+    if (!Validator.isNullOrEmpty(detail.repairStatus)) {
+      widgetList.add(_column(l10n.repairStatus, detail.repairStatus!, theme));
+    }
+    if (!Validator.isListNullOrEmpty(detail.channelList)) {
+      widgetList.add(_column(l10n.channelName, detail.channelList!.join(" | "), theme));
+    }
+    if (detail.stockAge != null) {
+      widgetList.add(_column(l10n.stockAge, detail.stockAge.toString(), theme));
+    }
+    if (!Validator.isNullOrEmpty(detail.lotName)) {
+      widgetList.add(_column(l10n.lotName, detail.lotName.toString(), theme));
+    }
+    if (!Validator.isNullOrEmpty(detail.otexSource)) {
+      widgetList.add(_column(l10n.otexSource, detail.otexSource.toString(), theme));
+    }
+
+    List<Widget> formatedList = [];
+    for (var i = 0; i < widgetList.length; i += 2) {
+      if (i > widgetList.length) {
+        break;
+      }
+      if ((i + 1) < widgetList.length) {
+        formatedList.add(Padding(
+          padding: const EdgeInsets.symmetric(vertical: Dimens.space_4),
+          child: Row(
+            children: [
+              Flexible(flex: 3, fit: FlexFit.tight, child: widgetList[i]),
+              Flexible(flex: 1, fit: FlexFit.tight, child: widgetList[i + 1]),
+            ],
+          ),
+        ));
+      } else {
+        formatedList.add(Padding(
+          padding: const EdgeInsets.symmetric(vertical: Dimens.space_4),
+          child: widgetList[i],
+        ));
+      }
+    }
+    return formatedList;
+  }
+
   _onDeviceScanned() {
     CshMlScannerUtil().openScanner(
       context,
@@ -99,16 +181,33 @@ class _DeviceDetailsWidgetState extends State<DeviceDetailsWidget> {
     );
   }
 
-  _row(String title, String value, ThemeData theme) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: Dimens.space_4),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Flexible(flex: 2, fit: FlexFit.tight, child: Text(title, style: theme.textTheme.titleSmall)),
-          Flexible(flex: 3, fit: FlexFit.tight, child: Text(value, style: theme.primaryTextTheme.titleMedium)),
-        ],
-      ),
+  _column(String title, String value, ThemeData theme) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text(title, style: theme.textTheme.bodySmall),
+        Text(value, style: theme.primaryTextTheme.headlineMedium),
+      ],
     );
+  }
+
+  Future<List<Step>> _getStockMovement(String deviceBarcode) {
+    var completer = Completer<List<Step>>();
+    DeviceDetailService.getDeviceStockMovement(deviceBarcode).listen((event) {
+      if (!Validator.isListNullOrEmpty(event?.stockMovementList)) {
+        List<Step> stepperList = event!.stockMovementList!.map((e) {
+          return Step(
+            title: CshTextNew.subTitle1(e.status ?? ""),
+            subtitle: CshTextNew.bodyText2(e.createdBy ?? ""),
+            content: CshTextNew.bodyText2(e.createdBy ?? ""),
+            isActive: e.isCurrentStatus ?? false,
+          );
+        }).toList();
+        completer.complete(stepperList);
+      }
+    }, onError: (error) {
+      completer.completeError(ApiErrorHelper.getErrorMessage(error).toString());
+    });
+    return completer.future;
   }
 }